requirements.txt
tensorflow
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
pillow
kagglehub
opencv-python
